+++ NEWS.md
@@ -1,5 +1,11 @@
 # RFishBC 0.0.13.9000 ongoing
 * `iSnap2Transect()`: Modified. Changed to handle bugs related to perfectly vertical or perfectly horizontal transects. Addresses [#27](https://github.com/droglenc/RFishBC/issues/27) (Thanks to Ben Neely).
+* `digitizeRadii()`: Modified. Removed `sepWindow()`. Added `deviceType=`.
+* `findScalingFactor()`: Modified. Removed `sepWindow()`. Added `deviceType=`.
+* `getID()`: Modified. Added `IDreplace=`. Added more examples of use of `IDpattern=` and `IDreplace=` and more tests.
+* `iGetimage()`: Modified. Removed `sepWindow=`. Added `deviceType=`.
+* `RFBCoptions()`: Modified. Removed `sepWindow()`. Added `deviceType=` and `IDreplace=`.
+* `showDigitizedImage()`: Modified. Removed `closeWindow=` and `sepWindow()`.
 
 # RFishBC 0.0.13 6-Oct-18
 * `combineData()`: Modified. Changed so that an age-0 fish with all plus-growth is still included in the returned data.frame when `deletePlusGrowth=TRUE` (the `ann` and `rad` variables will both be `NA`). Thanks to Ben Neely.